large refactor, tests pass, examples broken
